+++ packages/ui/src/components/profile-picture.tsx
