+++ packages/ui/src/components/search-bar.tsx
